# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracking API",
        "description": "Documentación de la Tracking API",
        "contact": {
            "email": "soporte@tudominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Mostrar perfil",
                "operationId": "5b219022e1f6d72ffff3925038abb15f",
                "responses": {
                    "200": {
                        "description": "Perfil obtenido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Actualizar perfil",
                "operationId": "6bd8da18a97d6b740678958d39e0dc40",
                "requestBody": {
                    "$ref": "#/components/schemas/UpdateProfileRequest"
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Cambiar contraseña",
                "operationId": "41230f5767a1259d34c33a2e4b481dd4",
                "requestBody": {
                    "$ref": "#/components/schemas/ChangePasswordRequest"
                },
                "responses": {
                    "200": {
                        "description": "Contraseña cambiada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de contraseña",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/notifications/email/enable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Habilitar Notificaciones por Email",
                "operationId": "c68a618e0f44c085ee8c1a2efefd6d6a",
                "responses": {
                    "200": {
                        "description": "Notificaciones por email habilitadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/notifications/email/disable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Deshabilitar Notificaciones por Email",
                "operationId": "6272d28efa603fc1f7531d5b3e6db878",
                "responses": {
                    "200": {
                        "description": "Notificaciones por email deshabilitadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/notifications/push/enable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Habilitar Notificaciones Push",
                "operationId": "e0a1b7d3f9d20470e9b0c462b6a05fd5",
                "responses": {
                    "200": {
                        "description": "Notificaciones Push habilitadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/notifications/push/disable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Deshabilitar Notificaciones Push",
                "operationId": "b0491cc470d121cf10ea6bcd882603e7",
                "responses": {
                    "200": {
                        "description": "Notificaciones Push deshabilitadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/newsletter/enable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Habilitar newsletter",
                "operationId": "c179199756da0ba68d0c6acc54bd4537",
                "responses": {
                    "200": {
                        "description": "Newsletter habilitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/newsletter/disable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Deshabilitar newsletter",
                "operationId": "80c6663bb7680fbfee295bb82d25597a",
                "responses": {
                    "200": {
                        "description": "Newsletter deshabilitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/2fa/enable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Habilitar 2FA",
                "operationId": "ef76011266d13fcff888e46b3b205cdf",
                "responses": {
                    "200": {
                        "description": "2FA habilitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/2fa/disable": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Deshabilitar 2FA",
                "operationId": "fa5ce3ee20dbb34f79b1e813b13621d9",
                "responses": {
                    "200": {
                        "description": "2FA deshabilitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/profile/private": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Deshabilitar perfil público",
                "operationId": "248da327df4d15194c5d07f273869d14",
                "responses": {
                    "200": {
                        "description": "Perfil público deshabilitado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/profile/public": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Habilitar perfil público",
                "operationId": "904c426bf1752f1e42b50cb8f157eff3",
                "responses": {
                    "200": {
                        "description": "Perfil público habilitado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/online-status/show": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Mostrar estado online",
                "operationId": "ea26c99b0e63b4658e3992c16e51a103",
                "responses": {
                    "200": {
                        "description": "Estado online visible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/online-status/hide": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Ocultar estado online",
                "operationId": "9b13276ec76f96608fe7eb2e0802dcb5",
                "responses": {
                    "200": {
                        "description": "Estado online oculto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Autenticación exitosa"
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johnd"
                                    },
                                    "photoUrl": {
                                        "type": "string",
                                        "example": "https://…"
                                    }
                                },
                                "type": "object"
                            },
                            "access_token": {
                                "type": "string",
                                "example": "jwt.token.here"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "bearer"
                            },
                            "expires_in": {
                                "type": "integer",
                                "example": 3600
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "GenericResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Operación exitosa"
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johnd"
                                    },
                                    "photoUrl": {
                                        "type": "string",
                                        "example": "https://…"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ChangePasswordRequest": {
                "required": [
                    "current_password",
                    "new_password"
                ],
                "properties": {
                    "current_password": {
                        "type": "string",
                        "format": "password",
                        "example": "oldPass123"
                    },
                    "new_password": {
                        "type": "string",
                        "format": "password",
                        "example": "newSecret456"
                    }
                },
                "type": "object"
            },
...
            "UpdateProfileRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "username": {
                        "type": "string",
                        "example": "juanperez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@ejemplo.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+34 123 456 789"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Desarrollador Full Stack..."
                    },
                    "location": {
                        "type": "string",
                        "example": "Madrid, España"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}